#10 slurm ability to get logs for specific job (#15)
* moved the obtaining of job logs to a separate provider class

* apply ability to use a relative log path to the jobLog provider, a few improvements of build.gradle

* minor changes

* changed jobId to long into the JobLogProvider

* moved the log directory path obtaining to log provider

* added a note to the getLogFilePath method

+++ src/main/java/com/epam/grid/engine/controller/job/JobOperationController.java
@@ -158,7 +158,7 @@ public class JobOperationController extends AbstractRestController {
             @ApiResponse(code = 404, message = NOT_FOUND),
             @ApiResponse(code = 500, message = INTERNAL_ERROR),
     })
-    public JobLogInfo getJobLogInfo(@PathVariable(JOB_ID) final int jobId,
+    public JobLogInfo getJobLogInfo(@PathVariable(JOB_ID) final long jobId,
                                     @RequestParam(value = LOG_TYPE_ID,
                                             required = false, defaultValue = "ERR") final JobLogInfo.Type logType,
                                     @RequestParam(value = "lines",
@@ -183,7 +183,7 @@ public class JobOperationController extends AbstractRestController {
             @ApiResponse(code = 404, message = NOT_FOUND)
     })
     public void getJobLogFile(final HttpServletResponse response,
-                              @PathVariable(JOB_ID) final int jobId,
+                              @PathVariable(JOB_ID) final long jobId,
                               @RequestParam(LOG_TYPE_ID) final JobLogInfo.Type logType) {
         try (InputStream inputStream = providerService.getJobLogFile(jobId, logType)) {
             writeStreamToResponse(response, inputStream, String.format("%d.%s", jobId, logType.getSuffix()));
